+++ tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/accessibility.tsx
@@ -48,3 +48,13 @@ export const stair_indicator: FeatureToggle = {
     `,
   component: CheckboxInput,
 };
+
+export const twelve_hour: FeatureToggle = {
+  name: 'Twelve-Hour Clock',
+  category: 'ACCESSIBILITY',
+  description: `
+      When toggled, will replace many instances of real-world time with AM/PM instead.
+    `,
+  component: CheckboxInput,
+};
+
